refactor: GameNumberGenerator.js Model로 이동
