+++ src/model/index.js
@@ -1,4 +1,4 @@
-import GameNumberGenerator from '../../utils/GameNumberGenerator.js';
+import GameNumberGenerator from './GameNumberGenerator.js';
 import GameNumberGeneratorError from '../../utils/error/GameNumberGeneratorError.js';
 import Validators from '../../utils/validator/index.js';
 import { ERROR_MESSAGE } from '../constants/Messages.js';
